Column list creation moved into separate file
