+++ c/expr/j_node.cc
@@ -20,6 +20,7 @@
 // IN THE SOFTWARE.
 //------------------------------------------------------------------------------
 #include "expr/base_expr.h"
+#include "expr/collist.h"
 #include "expr/j_node.h"
 #include "expr/workframe.h"   // dt::workframe
 namespace dt {
@@ -106,7 +107,9 @@ DataTable* allcols_jn::select(workframe& wf) {
     cols.reserve_extra(dti->ncols - j0);
     if (wf.nframes() > 1) {
       const strvec& dti_names = dti->get_names();
-      names.insert(names.end(), dti_names.begin() + j0, dti_names.end());
+      names.insert(names.end(),
+                   dti_names.begin() + static_cast<long>(j0),
+                   dti_names.end());
     }
 
     for (size_t j = j0; j < dti->ncols; ++j) {
@@ -165,14 +168,14 @@ class collist_jn : public j_node {
     strvec names;
 
   public:
-    collist_jn(std::vector<size_t>&& cols, strvec&& names_);
+    collist_jn(cols_intlist*);
     DataTable* select(workframe& wf) override;
     void delete_(workframe&) override;
 };
 
 
-collist_jn::collist_jn(std::vector<size_t>&& cols, strvec&& names_)
-  : indices(std::move(cols)), names(std::move(names_))
+collist_jn::collist_jn(cols_intlist* x)
+  : indices(std::move(x->indices)), names(std::move(x->names))
 {
   xassert(names.empty() || names.size() == indices.size());
 }
@@ -222,14 +225,14 @@ class exprlist_jn : public j_node {
     strvec names;
 
   public:
-    exprlist_jn(exprvec&& cols, strvec&& names_);
+    exprlist_jn(cols_exprlist*);
     DataTable* select(workframe& wf) override;
     void delete_(workframe&) override;
 };
 
 
-exprlist_jn::exprlist_jn(exprvec&& cols, strvec&& names_)
-  : exprs(std::move(cols)), names(std::move(names_))
+exprlist_jn::exprlist_jn(cols_exprlist* x)
+  : exprs(std::move(x->exprs)), names(std::move(x->names))
 {
   xassert(names.empty() || names.size() == exprs.size());
 }
@@ -273,313 +276,23 @@ void exprlist_jn::delete_(workframe&) {
 
 
 //------------------------------------------------------------------------------
-// j_node factory functions
-//------------------------------------------------------------------------------
-using intvec = std::vector<size_t>;
-using stypevec = std::vector<SType>;
-
-static const strvec typenames =
-  {"?", "boolean", "integer", "string", "expr", "type"};
-
-static stypevec stBOOL = {SType::BOOL};
-static stypevec stINT = {SType::INT8, SType::INT16, SType::INT32, SType::INT64};
-static stypevec stFLOAT = {SType::FLOAT32, SType::FLOAT64};
-static stypevec stSTR = {SType::STR32, SType::STR64};
-static stypevec stOBJ = {SType::OBJ};
-
-
-class jnode_maker
-{
-  private:
-    enum class list_type : size_t {
-      UNKNOWN, BOOL, INT, STR, EXPR, TYPE
-    };
-
-    workframe& wf;
-    const DataTable* dt0;
-    list_type type;
-    intvec  indices;
-    exprvec exprs;
-    strvec  names;
-    size_t  k;  // The index of the current element
-    bool is_update;
-    bool is_delete;
-    bool has_new_columns;
-    size_t : 40;
-
-  public:
-    jnode_maker(py::robj src, workframe& wf_) : wf(wf_)
-    {
-      is_update = (wf.get_mode() == EvalMode::UPDATE);
-      is_delete = (wf.get_mode() == EvalMode::DELETE);
-      dt0 = wf.get_datatable(0);
-      type = list_type::UNKNOWN;
-      k = 0;
-
-      if (is_PyBaseExpr(src))   _process_element_expr(src);
-      else if (src.is_int())    _process_element_int(src);
-      else if (src.is_string()) _process_element_string(src);
-      else if (src.is_slice())  _process_element_slice(src);
-      else if (src.is_type())   _process_element_type(src);
-      else if (src.is_ltype())  _process_element_ltype(src);
-      else if (src.is_stype())  _process_element_stype(src);
-
-      else if (src.is_list_or_tuple()) {
-        py::olist srclist = src.to_pylist();
-        size_t nelems = srclist.size();
-        for (size_t i = 0; i < nelems; ++i) {
-          _process_element(srclist[i]);
-        }
-      }
-      else if (src.is_dict()) {
-        if (is_delete) {
-          throw TypeError() << "When del operator is applied, the `j` selector "
-              "cannot be a dictionary";
-        }
-        type = list_type::EXPR;
-        for (auto kv : src.to_pydict()) {
-          if (!kv.first.is_string()) {
-            throw TypeError()
-                << "Keys in the `j` selector dictionary must be strings";
-          }
-          names.push_back(kv.first.to_string());
-          _process_element(kv.second);
-        }
-      }
-      else if (src.is_iterable() && !src.is_bytes()) {
-        for (auto elem : src.to_oiter()) {
-          _process_element(elem);
-        }
-      }
-      else {
-        throw TypeError()
-          << "Unsupported `j` selector of type " << src.typeobj();
-      }
-    }
-
-    j_node* get_node() {
-      if (type == list_type::BOOL && k != dt0->ncols) {
-        throw ValueError()
-            << "The length of boolean list in `j` does not match the "
-               "number of columns in the Frame: "
-            << k << " vs " << dt0->ncols;
-      }
-      // A list of "EXPR" type may be either a list of plain column selectors
-      // (such as `f.A`), or a list of more complicated expressions. In the
-      // former case the vector of `indices` will be the same size as `exprs`,
-      // and we return a `collist_jn` node. In the latter case, the
-      // `exprlist_jn` node is created.
-      if (exprs.size() > indices.size()) {
-        xassert(type == list_type::EXPR);
-        return new exprlist_jn(std::move(exprs), std::move(names));
-      }
-      return new collist_jn(std::move(indices), std::move(names));
-    }
-
-  private:
-    void _set_type(list_type t) {
-      if (type == list_type::UNKNOWN) type = t;
-      if (type == t) return;
-      if (k) {
-        throw TypeError()
-            << "Mixed selector types in `j` are not allowed. Element "
-            << k << " is of type "
-            << typenames[static_cast<size_t>(t)]
-            << ", whereas the previous element(s) were of type "
-            << typenames[static_cast<size_t>(type)];
-      } else {
-        throw TypeError()
-            << "The values in the `j` selector dictionary must be "
-               "expressions, not "
-            << typenames[static_cast<size_t>(t)] << 's';
-      }
-    }
-
-    void _process_element(py::robj elem) {
-      if (is_PyBaseExpr(elem))   _process_element_expr(elem);
-      else if (elem.is_int())    _process_element_int(elem);
-      else if (elem.is_bool())   _process_element_bool(elem);
-      else if (elem.is_string()) _process_element_string(elem);
-      else if (elem.is_slice())  _process_element_slice(elem);
-      else if (elem.is_type())   _process_element_type(elem);
-      else if (elem.is_ltype())  _process_element_ltype(elem);
-      else if (elem.is_stype())  _process_element_stype(elem);
-      else {
-        throw TypeError()
-            << "Element " << k << " in the `j` selector list has type `"
-            << elem.typeobj() << "`, which is not supported";
-      }
-      ++k;
-    }
-
-    void _process_element_int(py::robj elem) {
-      _set_type(list_type::INT);
-      int64_t i = elem.to_int64_strict();
-      int64_t icols = static_cast<int64_t>(dt0->ncols);
-      if (i < -icols || i >= icols) {
-        throw ValueError()
-            << "Column index `" << i << "` is invalid for a Frame with "
-            << icols << " column" << (icols == 1? "" : "s");
-      }
-      if (i < 0) i += icols;
-      indices.push_back(static_cast<size_t>(i));
-    }
-
-    void _process_element_bool(py::robj elem) {
-      _set_type(list_type::BOOL);
-      int8_t t = elem.to_bool_strict();
-      if (t) {
-        indices.push_back(k);
-      }
-    }
-
-    void _process_element_string(py::robj elem) {
-      _set_type(list_type::STR);
-      if (is_update) {
-        int64_t j = dt0->colindex(elem);
-        if (j == -1) {
-          has_new_columns = true;
-          names.resize(indices.size());
-          names.push_back(elem.to_string());
-        }
-        indices.push_back(static_cast<size_t>(j));
-      } else {
-        size_t j = dt0->xcolindex(elem);
-        indices.push_back(j);
-      }
-    }
-
-    void _process_element_expr(py::robj elem) {
-      _set_type(list_type::EXPR);
-      py::oobj res = elem.invoke("_core");
-      xassert(res.typeobj() == &py::base_expr::Type::type);
-      auto pybe = reinterpret_cast<py::base_expr*>(res.to_borrowed_ref());
-      dt::base_expr* pexpr = pybe->release();
-      auto expr = std::unique_ptr<dt::base_expr>(pexpr);
-
-      dt::expr_column* colexpr = dynamic_cast<dt::expr_column*>(pexpr);
-      if (colexpr) {
-        size_t frid = colexpr->get_frame_id();
-        if (frid == 0) {
-          size_t i = colexpr->get_col_index(wf);
-          indices.push_back(i);
-        }
-        else if (frid >= wf.nframes()) {
-          throw ValueError() << "Item " << k << " of the `j` selector list "
-              "references a non-existing join frame";
-        }
-      }
-      exprs.push_back(std::move(expr));
-    }
-
-    void _process_element_slice(py::robj elem) {
-      py::oslice ssrc = elem.to_oslice();
-      if (ssrc.is_numeric()) return _process_element_numslice(ssrc);
-      if (ssrc.is_string())  return _process_element_strslice(ssrc);
-      throw TypeError() << ssrc << " is neither integer- nor string-valued";
-    }
-
-    void _process_element_numslice(py::oslice ssrc) {
-      _set_type(list_type::INT);
-      size_t len = dt0->ncols;
-      size_t start, count, step;
-      ssrc.normalize(len, &start, &count, &step);
-      indices.reserve(indices.size() + count);
-      for (size_t i = 0; i < count; ++i) {
-        indices.push_back(start + i * step);
-      }
-    }
-
-    void _process_element_strslice(py::oslice ssrc) {
-      _set_type(list_type::STR);
-      py::oobj ostart = ssrc.start_obj();
-      py::oobj ostop = ssrc.stop_obj();
-      size_t start = ostart.is_none()? 0 : dt0->xcolindex(ostart);
-      size_t end = ostop.is_none()? dt0->ncols - 1 : dt0->xcolindex(ostop);
-      size_t len = start <= end? end - start + 1 : start - end + 1;
-      indices.reserve(len);
-      if (start <= end) {
-        for (size_t i = start; i <= end; ++i) {
-          indices.push_back(i);
-        }
-      } else {
-        // Careful with the case when `end = 0`. In this case a regular for-loop
-        // `(i = start; i >= end; --i)` becomes infinite.
-        size_t i = start;
-        do {
-          indices.push_back(i);
-        } while (i-- != end);
-      }
-    }
-
-    void _process_element_type(py::robj elem) {
-      _set_type(list_type::TYPE);
-      PyTypeObject* et = reinterpret_cast<PyTypeObject*>(elem.to_borrowed_ref());
-      if (et == &PyLong_Type)            _select_types(stINT);
-      else if (et == &PyFloat_Type)      _select_types(stFLOAT);
-      else if (et == &PyUnicode_Type)    _select_types(stSTR);
-      else if (et == &PyBool_Type)       _select_types(stBOOL);
-      else if (et == &PyBaseObject_Type) _select_types(stOBJ);
-      else {
-        throw ValueError()
-            << "Unknown type " << elem << " used as a `j` selector";
-      }
-    }
-
-    void _process_element_ltype(py::robj elem) {
-      _set_type(list_type::TYPE);
-      size_t lt = elem.get_attr("value").to_size_t();
-      switch (static_cast<LType>(lt)) {
-        case LType::BOOL:   _select_types(stBOOL); break;
-        case LType::INT:    _select_types(stINT); break;
-        case LType::REAL:   _select_types(stFLOAT); break;
-        case LType::STRING: _select_types(stSTR); break;
-        case LType::OBJECT: _select_types(stOBJ); break;
-        default:
-          throw TypeError() << "Unknown ltype value " << lt;
-      }
-    }
-
-    void _process_element_stype(py::robj elem) {
-      _set_type(list_type::TYPE);
-      size_t st = elem.get_attr("value").to_size_t();
-      _select_types({static_cast<SType>(st)});
-    }
-
-    void _select_types(const std::vector<SType>& stypes) {
-      size_t ncols = dt0->ncols;
-      for (size_t i = 0; i < ncols; ++i) {
-        SType st = dt0->columns[i]->stype();
-        for (SType s : stypes) {
-          if (s == st) {
-            indices.push_back(i);
-            break;
-          }
-        }
-      }
-    }
-};
-
-
-static j_node* _make(py::robj src, workframe& wf) {
-  // The most common case is ":", a trivial slice
-  if ((src.is_slice() && src.to_oslice().is_trivial())
-      || src.is_none() || src.is_ellipsis()) {
-    return new allcols_jn();
-  }
-  return jnode_maker(src, wf).get_node();
-}
-
-
-
-
-//------------------------------------------------------------------------------
 // j_node
 //------------------------------------------------------------------------------
 
 jptr j_node::make(py::robj src, workframe& wf) {
-  return jptr(_make(src, wf));
+  // The most common case is ":", a trivial slice
+  if ((src.is_slice() && src.to_oslice().is_trivial())
+      || src.is_none() || src.is_ellipsis()) {
+    return jptr(new allcols_jn());
+  }
+  collist_ptr cl = make_collist(src, wf);
+  auto cl_int = dynamic_cast<cols_intlist*>(cl.get());
+  auto cl_expr = dynamic_cast<cols_exprlist*>(cl.get());
+  xassert(cl_int || cl_expr);
+  return cl_int? jptr(new collist_jn(cl_int))
+               : jptr(new exprlist_jn(cl_expr));
 }
+
 
 j_node::~j_node() {}
